requirements.txt
cffi==1.14.5
cryptography==3.4.7
dj-config-url==0.1.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.19
django-appconf==1.0.4
django-cryptography==1.0
django-filter==2.4.0
django-session-csrf==0.7.1
django-stdimage==5.3.0
djangorestframework==3.12.4
greenlet==1.0.0
gunicorn==20.0.4
home==0.0.3
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
pyserial==3.5
pytz==2021.1
rfxcom==0.5.0
SQLAlchemy==1.4.7
sqlparse==0.4.1
static3==0.7.0
uuid==1.30
